Move hash_index into index subdirectory
