+++ libvast/CMakeLists.txt
@@ -340,12 +340,12 @@ set(tests
     test/format/writer.cpp
     test/format/zeek.cpp
     test/hash.cpp
-    test/hash_index.cpp
     test/http.cpp
     test/ids.cpp
     test/index/address_index.cpp
     test/index/arithmetic_index.cpp
     test/index/enumeration_index.cpp
+    test/index/hash_index.cpp
     test/index/list_index.cpp
     test/index/port_index.cpp
     test/index/string_index.cpp
