+++ libvast/src/value_index_factory.cpp
@@ -18,10 +18,10 @@
 #include "vast/concept/parseable/vast/base.hpp"
 #include "vast/detail/bit.hpp"
 #include "vast/detail/type_traits.hpp"
-#include "vast/hash_index.hpp"
 #include "vast/index/address_index.hpp"
 #include "vast/index/arithmetic_index.hpp"
 #include "vast/index/enumeration_index.hpp"
+#include "vast/index/hash_index.hpp"
 #include "vast/index/list_index.hpp"
 #include "vast/index/port_index.hpp"
 #include "vast/index/string_index.hpp"
