+++ libvast/vast/index/hash_index.hpp
